pyproject.toml
[tool.poetry]
name = "sql-generator"
version = "0.1.0"
description = ""
authors = ["Blake Enyart <blake@mashey.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Jinja2 = "^2.11.3"
snowflake-sqlalchemy = "^1.2.4"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.4b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
